move the test utils to a separate module
